pyproject.toml
[project]
name = "phasmo-update-bot"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "dodaucy", email = "dodaucy@gmail.com"},
]
dependencies = ["httpx>=0.27.2", "beautifulsoup4>=4.12.3", "pydantic>=2.9.2", "python-dotenv>=1.0.1", "databases[aiomysql]>=0.9.0"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
